singlevm: Split up and clean up legacy parts

+++ testutil/singlevm/run_controller.sh
@@ -1,5 +1,17 @@
 #!/bin/bash
 ciao_host=$(hostname)
+ciao_dir=/var/lib/ciao
+ciao_data_dir=${ciao_dir}/data
+ciao_ctl_dir=${ciao_data_dir}/controller
+
+#Create controller dirs
+echo "Making ciao workloads dir: ${ciao_ctl_dir}/workloads"
+sudo mkdir -p ${ciao_ctl_dir}/workloads
+if [ ! -d ${ciao_ctl_dir}/workloads ]
+then
+	echo "FATAL ERROR: Unable to create ${ciao_ctl_dir}/workloads}"
+	exit 1
+fi
 
 sudo rm ciao-controller.db-shm ciao-controller.db-wal ciao-controller.db /tmp/ciao-controller-stats.db
